types

+++ app/Report/ReportHtml.php
@@ -383,16 +383,16 @@ class ReportHtml extends ReportBase
      * Create an image.
      *
      * @param string $file
-     * @param float  $x
-     * @param float  $y
-     * @param float  $w
-     * @param float  $h
+     * @param int       $x
+     * @param int       $y
+     * @param int       $w
+     * @param int       $h
      * @param string $align
      * @param string $ln
      *
      * @return ReportHtmlImage
      */
-    public function createImage(string $file, float $x, float $y, float $w, float $h, string $align, string $ln): ReportHtmlImage
+    public function createImage(string $file, int $x, int $y, int $w, int $h, string $align, string $ln): ReportHtmlImage
     {
         return new ReportHtmlImage($file, $x, $y, $w, $h, $align, $ln);
     }
